refactor: split mcp_bridge, mcp_config, and genies into submodules (partial #93)
Split the three most oversized Rust files into focused submodules:

- mcp_bridge.rs (784 lines) -> mcp_bridge/{mod,types,state,server,commands}.rs
- mcp_config.rs (615 lines) -> mcp_config/{mod,types,providers,config_io,commands}.rs
- genies.rs (578 lines) -> genies/{mod,types,scanning,parsing,install,commands,tests}.rs

All public APIs are re-exported from mod.rs so external callers are
unaffected. Tauri command registrations updated to use nested paths.
All 13 genies tests pass. cargo check clean with no warnings.
