+++ src-tauri/src/lib.rs
@@ -161,12 +161,12 @@ pub fn run() {
             mcp_server::mcp_sidecar_health,
             mcp_server::mcp_bridge_client_count,
             mcp_server::mcp_bridge_connected_clients,
-            mcp_bridge::mcp_bridge_respond,
-            mcp_config::mcp_config_get_status,
-            mcp_config::mcp_config_diagnose,
-            mcp_config::mcp_config_preview,
-            mcp_config::mcp_config_install,
-            mcp_config::mcp_config_uninstall,
+            mcp_bridge::commands::mcp_bridge_respond,
+            mcp_config::commands::mcp_config_get_status,
+            mcp_config::commands::mcp_config_diagnose,
+            mcp_config::commands::mcp_config_preview,
+            mcp_config::commands::mcp_config_install,
+            mcp_config::commands::mcp_config_uninstall,
             hot_exit::commands::hot_exit_capture,
             hot_exit::commands::hot_exit_restore,
             hot_exit::commands::hot_exit_inspect_session,
@@ -181,9 +181,9 @@ pub fn run() {
             tab_transfer::focus_existing_window,
             tab_transfer::remove_tab_from_window,
             get_default_shell,
-            genies::get_genies_dir,
-            genies::list_genies,
-            genies::read_genie,
+            genies::commands::get_genies_dir,
+            genies::commands::list_genies,
+            genies::commands::read_genie,
             ai_provider::detect_ai_providers,
             ai_provider::run_ai_prompt,
             ai_provider::read_env_api_keys,
