+++ src-tauri/src/mcp_bridge/server.rs
@@ -1,295 +1,23 @@
-//! MCP Bridge - WebSocket server for AI assistant communication.
+//! MCP Bridge WebSocket server loop and connection handling.
 //!
-//! Provides a WebSocket server that MCP sidecars connect to.
-//! Access model:
-//! - Read operations: All clients can execute simultaneously
-//! - Write operations: Serialized via write lock, released after each write
-//!
-//! Port discovery:
-//! - Server binds to port 0 (OS assigns available port)
-//! - Actual port written to Tauri's app data directory (platform-specific)
-//! - MCP sidecar uses platform-specific path to find the app data directory
+//! Manages the TCP listener, WebSocket upgrades, per-client message loops,
+//! and request routing to the frontend.
 
-use crate::app_paths;
+use super::state::{
+    get_bridge_state, get_shutdown_holder, get_write_lock, is_read_only_operation,
+    remove_port_file, write_port_file, ClientConnection, PendingRequest,
+};
+use super::types::{
+    ClientIdentity, McpRequest, McpRequestEvent, McpResponse, WsMessage,
+};
 use futures_util::{SinkExt, StreamExt};
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-use std::fs;
 use std::net::SocketAddr;
-use std::sync::Arc;
 use std::time::Instant;
 use tauri::AppHandle;
 use tauri::Emitter;
 use tokio::net::{TcpListener, TcpStream};
-use tokio::sync::{mpsc, oneshot, Mutex, RwLock};
+use tokio::sync::{mpsc, oneshot};
 use tokio_tungstenite::{accept_async, tungstenite::Message};
-
-/// Message format for WebSocket communication with the sidecar.
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct WsMessage {
-    pub id: String,
-    #[serde(rename = "type")]
-    pub msg_type: String,
-    pub payload: serde_json::Value,
-}
-
-/// MCP request from the sidecar.
-#[derive(Clone, Debug)]
-pub struct McpRequest {
-    pub request_type: String,
-    pub args: serde_json::Value,
-}
-
-impl McpRequest {
-    fn from_value(value: serde_json::Value) -> Result<Self, String> {
-        let obj = value.as_object().ok_or("Request must be an object")?;
-
-        let request_type = obj
-            .get("type")
-            .and_then(|v| v.as_str())
-            .ok_or("Request must have a 'type' field")?
-            .to_string();
-
-        let mut args = serde_json::Map::new();
-        for (key, val) in obj.iter() {
-            if key != "type" {
-                args.insert(key.clone(), val.clone());
-            }
-        }
-
-        Ok(McpRequest {
-            request_type,
-            args: serde_json::Value::Object(args),
-        })
-    }
-}
-
-/// MCP response to send back to the sidecar.
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct McpResponse {
-    pub success: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub data: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub error: Option<String>,
-}
-
-/// Event payload sent to frontend.
-/// Note: args is serialized as a JSON string to avoid Tauri IPC double-encoding issues
-/// with nested serde_json::Value. The frontend must parse this string.
-#[derive(Clone, Debug, Serialize)]
-pub struct McpRequestEvent {
-    pub id: String,
-    #[serde(rename = "type")]
-    pub request_type: String,
-    /// JSON-serialized args string (frontend must JSON.parse this)
-    pub args_json: String,
-}
-
-/// Response from frontend via command.
-#[derive(Clone, Debug, Deserialize)]
-pub struct McpResponsePayload {
-    pub id: String,
-    pub success: bool,
-    pub data: Option<serde_json::Value>,
-    pub error: Option<String>,
-}
-
-/// Connected client info exposed to the frontend.
-#[derive(Clone, Debug, Serialize)]
-pub struct ConnectedClientInfo {
-    pub name: String,
-    pub version: Option<String>,
-}
-
-/// Client identity information sent during handshake.
-#[derive(Clone, Debug, Default, serde::Deserialize)]
-struct ClientIdentity {
-    /// Client name (e.g., "claude-code", "codex-cli", "cursor")
-    name: String,
-    /// Client version
-    #[serde(default)]
-    version: Option<String>,
-    /// Process ID
-    #[serde(default)]
-    #[allow(dead_code)]
-    pid: Option<u32>,
-    /// Parent process name
-    #[serde(rename = "parentProcess")]
-    #[serde(default)]
-    #[allow(dead_code)]
-    parent_process: Option<String>,
-}
-
-impl ClientIdentity {
-    /// Get display name for logging (debug only).
-    #[cfg(debug_assertions)]
-    fn display_name(&self) -> String {
-        if let Some(ref version) = self.version {
-            format!("{} v{}", self.name, version)
-        } else {
-            self.name.clone()
-        }
-    }
-}
-
-/// Connected client information.
-struct ClientConnection {
-    #[allow(dead_code)]
-    id: u64,
-    #[allow(dead_code)]
-    addr: SocketAddr,
-    tx: mpsc::UnboundedSender<String>,
-    shutdown: Option<oneshot::Sender<()>>,
-    #[allow(dead_code)]
-    connected_at: Instant,
-    /// Client identity (set after identify message)
-    identity: Option<ClientIdentity>,
-}
-
-/// Bridge state shared across connections.
-struct BridgeState {
-    /// All connected clients (equal access for reads).
-    clients: HashMap<u64, ClientConnection>,
-    /// Pending requests waiting for responses from frontend.
-    pending: HashMap<String, PendingRequest>,
-    /// Counter for generating unique client IDs.
-    next_client_id: u64,
-}
-
-/// Pending request with client ID for routing response.
-struct PendingRequest {
-    response_tx: oneshot::Sender<McpResponse>,
-    #[allow(dead_code)]
-    client_id: u64,
-}
-
-/// Global bridge state.
-static BRIDGE_STATE: std::sync::OnceLock<Arc<Mutex<BridgeState>>> = std::sync::OnceLock::new();
-
-/// Server shutdown signal.
-static SHUTDOWN_TX: std::sync::OnceLock<Arc<RwLock<Option<oneshot::Sender<()>>>>> =
-    std::sync::OnceLock::new();
-
-/// Write lock for serializing write operations.
-/// All clients can read simultaneously, but writes are serialized.
-static WRITE_LOCK: std::sync::OnceLock<Arc<tokio::sync::Mutex<()>>> = std::sync::OnceLock::new();
-
-fn get_bridge_state() -> Arc<Mutex<BridgeState>> {
-    BRIDGE_STATE
-        .get_or_init(|| {
-            Arc::new(Mutex::new(BridgeState {
-                clients: HashMap::new(),
-                pending: HashMap::new(),
-                next_client_id: 1,
-            }))
-        })
-        .clone()
-}
-
-fn get_shutdown_holder() -> Arc<RwLock<Option<oneshot::Sender<()>>>> {
-    SHUTDOWN_TX
-        .get_or_init(|| Arc::new(RwLock::new(None)))
-        .clone()
-}
-
-fn get_write_lock() -> Arc<tokio::sync::Mutex<()>> {
-    WRITE_LOCK
-        .get_or_init(|| Arc::new(tokio::sync::Mutex::new(())))
-        .clone()
-}
-
-/// Write the port to the port file for MCP sidecar discovery.
-/// Uses atomic write to prevent partial reads by the sidecar.
-fn write_port_file(app: &AppHandle, port: u16) -> Result<(), String> {
-    let path = app_paths::get_port_file_path(app)?;
-
-    // Create app data directory if it doesn't exist
-    if let Some(parent) = path.parent() {
-        fs::create_dir_all(parent).map_err(|e| {
-            format!(
-                "Failed to create app data directory {:?}: {}",
-                parent, e
-            )
-        })?;
-    }
-
-    // Write port atomically to prevent partial reads
-    app_paths::atomic_write_file(&path, port.to_string().as_bytes())?;
-
-    #[cfg(debug_assertions)]
-    eprintln!("[MCP Bridge] Port {} written to {:?}", port, path);
-
-    Ok(())
-}
-
-/// Remove the port file when bridge stops.
-/// Logs errors for non-NotFound failures (permission issues, etc.)
-fn remove_port_file(app: &AppHandle) {
-    match app_paths::get_port_file_path(app) {
-        Ok(path) => {
-            match fs::remove_file(&path) {
-                Ok(()) => {
-                    #[cfg(debug_assertions)]
-                    eprintln!("[MCP Bridge] Port file removed: {:?}", path);
-                }
-                Err(e) if e.kind() == std::io::ErrorKind::NotFound => {
-                    // Already removed - not an error
-                }
-                Err(e) => {
-                    // Real error - log it
-                    eprintln!(
-                        "[MCP Bridge] Warning: Failed to remove port file {:?}: {}",
-                        path, e
-                    );
-                }
-            }
-        }
-        Err(e) => {
-            eprintln!("[MCP Bridge] Warning: Cannot determine port file path: {}", e);
-        }
-    }
-}
-
-/// Check if an operation is read-only.
-fn is_read_only_operation(request_type: &str) -> bool {
-    matches!(
-        request_type,
-        // Document read operations
-        "document.getContent"
-            | "document.search"
-            // Selection/cursor read operations
-            | "selection.get"
-            | "cursor.getContext"
-            // Metadata operations
-            | "outline.get"
-            | "metadata.get"
-            // Window/workspace read operations
-            | "windows.list"
-            | "windows.getFocused"
-            | "workspace.getDocumentInfo"
-            | "workspace.listRecentFiles"
-            | "workspace.getInfo"
-            // Tab read operations
-            | "tabs.list"
-            | "tabs.getActive"
-            | "tabs.getInfo"
-            // Editor state operations
-            | "editor.getUndoState"
-            // Suggestion read operations
-            | "suggestion.list"
-            // Paragraph read operations
-            | "paragraph.read"
-            // Protocol/structure read operations
-            | "protocol.getCapabilities"
-            | "protocol.getRevision"
-            | "structure.getAst"
-            | "structure.getDigest"
-            | "structure.listBlocks"
-            | "structure.resolveTargets"
-            | "structure.getSection"
-    )
-}
 
 /// Start the MCP bridge WebSocket server.
 /// Returns the actual port the server is listening on.
@@ -719,66 +447,4 @@ async fn handle_message(text: &str, client_id: u64, app: &AppHandle) -> Result<(
         .map_err(|e| format!("Failed to send response: {}", e))?;
 
     Ok(())
-}
-
-/// Tauri command to send a response from the frontend.
-#[tauri::command]
-pub async fn mcp_bridge_respond(payload: McpResponsePayload) -> Result<(), String> {
-    let state = get_bridge_state();
-    let mut guard = state.lock().await;
-
-    #[cfg(debug_assertions)]
-    if guard.pending.contains_key(&payload.id) {
-        eprintln!("[MCP Bridge] Response received for {}", payload.id);
-    } else {
-        eprintln!(
-            "[MCP Bridge] Response for unknown/expired request {}",
-            payload.id
-        );
-    }
-
-    if let Some(pending) = guard.pending.remove(&payload.id) {
-        let response = McpResponse {
-            success: payload.success,
-            data: payload.data,
-            error: payload.error,
-        };
-        pending
-            .response_tx
-            .send(response)
-            .map_err(|_| "Response channel closed")?;
-    }
-
-    Ok(())
-}
-
-/// Check if the bridge has any connected clients.
-#[allow(dead_code)]
-pub async fn is_client_connected() -> bool {
-    let state = get_bridge_state();
-    let guard = state.lock().await;
-    !guard.clients.is_empty()
-}
-
-/// Get count of connected clients.
-pub async fn client_count() -> usize {
-    let state = get_bridge_state();
-    let guard = state.lock().await;
-    guard.clients.len()
-}
-
-/// Get list of connected clients with their identities.
-pub async fn connected_clients() -> Vec<ConnectedClientInfo> {
-    let state = get_bridge_state();
-    let guard = state.lock().await;
-    guard
-        .clients
-        .values()
-        .filter_map(|c| {
-            c.identity.as_ref().map(|i| ConnectedClientInfo {
-                name: i.name.clone(),
-                version: i.version.clone(),
-            })
-        })
-        .collect()
 }
